+++ CodeStarterCamp_Week1/iceCream.swift
@@ -1,0 +1,24 @@
+//
+//  main.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by yagom.
+//  Copyright © yagom academy. All rights reserved.
+//
+
+import Foundation
+
+func makeIceCream() {
+    for _ in 1...8 {
+        for _ in 1...11 {
+            print("*", terminator: "")
+        }
+        print("")
+    }
+}
+
+func makeStick() {
+    for _ in 1...4 {
+        print("    | |")
+    }
+}
